re-organize cyclegan code into more modules
