+++ external/fv3fit/fv3fit/pytorch/cyclegan/__init__.py
@@ -1,9 +1,10 @@
-from .train import train_autoencoder, AutoencoderHyperparameters
-from .network import GeneratorConfig, DiscriminatorConfig
+from .train_autoencoder import train_autoencoder, AutoencoderHyperparameters
 from .train_cyclegan import (
     train_cyclegan,
     CycleGANHyperparameters,
-    CycleGANNetworkConfig,
     CycleGANTrainingConfig,
-    CycleGAN,
 )
+from .discriminator import DiscriminatorConfig
+from .generator import GeneratorConfig
+from .cyclegan_trainer import CycleGANNetworkConfig
+from .reloadable import CycleGAN
