+++ external/fv3fit/fv3fit/pytorch/predict.py
@@ -113,14 +113,17 @@ class PytorchPredictor(Predictor):
         )
         # dimensions are [time, tile, x, y, z],
         # we must combine [time, tile] into one sample dimension
-        return torch.reshape(
+        reshaped = torch.reshape(
             packed, (packed.shape[0] * packed.shape[1],) + tuple(packed.shape[2:]),
         )
+        # torch expects channels before x, y so we have to transpose
+        transposed = reshaped.permute([0, 3, 1, 2])
+        return transposed
 
     def unpack_tensor(self, data: torch.Tensor) -> xr.Dataset:
         data = torch.reshape(data, (-1, 6) + tuple(data.shape[1:]))
         return _unpack_tensor(
-            data,
+            data.permute([0, 1, 3, 4, 2]),  # convert from channels (z) first to last
             varnames=tuple(str(item) for item in self.output_variables),
             scalers=self.scalers,
             dims=["time", "tile", "x", "y", "z"],
